scripts/requirements.txt
fire
orjson
requests
python-dotenv
tqdm